chore(build): Move Nix commit for releases to latest NixOS stable
Updates a few build dependencies, most notably Go to 1.11. The chosen
commit is the current stable `nixos-18.09` tag in nixpkgs-channels.

While doing this I also checked whether it is now possible to remove
the exception for allowing references, but it's not.

Go may in the future gain support for fully reproducible builds, see
for example issue #16860 on their compiler. At that point maybe we can
drop a few settings here.